requirements.txt
fastapi>=0.95,<0.110
uvicorn>=0.23,<0.31
pydantic>=1.10,<2.0
fastapi-utils>=0.2.1
pandas>=1.5
numpy>=1.24
scipy>=1.10
scikit-learn>=1.2
sklearn-pandas>=2.2.0
matplotlib>=3.7
seaborn>=0.12
statsmodels>=0.14
prophet>=1.1
PyYAML>=6.0
typing-inspect>=0.9.0
openpyxl>=3.1
pytest>=7.4
